anxiety-iosched: Standardize loop iterator data type
As previously declared in anxiety_data, sync_ratio is of size uint8_t.
Therefore, the loop iterator can never exceed the size of a uint8_t.
Correct for this inconsistency.

+++ block/anxiety-iosched.c
@@ -61,7 +61,7 @@ static inline int __anxiety_dispatch(struct request_queue *q,
 static int anxiety_dispatch(struct request_queue *q, int force)
 {
 	struct anxiety_data *adata = q->elevator->elevator_data;
-	int batched;
+	uint8_t batched;
 
 	/* Make sure we can even process any requests at all */
 	if (!anxiety_can_dispatch(adata))
